refactor: split survey route helpers
